Port is incorrectly written

+++ OnionProxy/src/main/kotlin/edu/ntnu/tobiasth/onionproxy/Config.kt
@@ -3,6 +3,6 @@ package edu.ntnu.tobiasth.onionproxy
 class Config {
     companion object {
         const val SOCKS_PORT: Int = 1080
-        const val SOCKS_VERSION: Byte = 5
+        const val SOCKS_VERSION: Int = 5
     }
 }
